Minimarts Problem
construction of minimarts along a stretch of road. The possible sites
for the minimarts are given by real numbers x1; :::; xn, each of which
specifies the position along the road measured in miles from its
beginning. Assume that the road is
a straight line. If you place a minimart at location xi, your company
will make a profit of ri > 0 dollars.Every pair of minimarts should be
at least 5 miles apart. You'd like to place
minimarts at a subset of the sites so as to maximize total prot,
subject to this restriction.
The input is given as a list of n pairs (x1; r1); :::; (xn; rn) where
the xi's are sorted in increasing
order.